update fmt clippy and wasm32
