+++ Cargo.toml
@@ -1,35 +1,38 @@
 [package]
 name = "ptv"
-version = "0.2.2"
-edition = "2021"
+version = "0.2.3"
+edition = "2024"
 license = "MIT"
 description = "A Rust library for the Public Transport Victoria (PTV) API"
 repository = "https://github.com/tascord/ptvrs"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[dependencies]
-anyhow = "1.0.81"
-chrono = { version = "0.4.35", features = ["serde"] }
-colored = "2.1.0"
-derive_more = "0.99.18"
+[target.'cfg(target_arch = "x86_64")'.dependencies]
+colored = "3.0.0"
 dotenv = "0.15.0"
 hex = "0.4.3"
 hmac = "0.12.1"
-itertools = "0.13.0"
 once_cell = "1.19.0"
 reqwest = { version = "0.12.0", features = ["json"] }
+sha1 = "0.10.6"
+tokio = { version = "1.36.0", features = ["full"] }
+
+[dependencies]
+anyhow = "1.0.81"
+chrono = { version = "0.4.35", features = ["serde"] }
+derive_more = { version = "2", features = ["display", "debug", "from"] }
+itertools = "0.14.0"
 serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.114"
-sha1 = "0.10.6"
-to_and_fro = "0.5.0"
-tokio = { version = "1.36.0", features = ["full"] }
+to_and_fro = "0.7.1"
 url-escape = "0.1.1"
 
 [dev-dependencies]
 futures = "0.3.30"
 ptvrs-macros = { path = "ptvrs-macros" }
 
-
 [workspace]
 members = ["ptvrs-macros"]
+
+[features]
